% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_kb.R
\name{find_kb_mac}
\alias{find_kb_mac}
\title{Finds shortcuts based on matched pattern on mac keybindings.}
\usage{
find_kb_mac(pattern, ignore_case = TRUE, data = shortcuts_table)
}
\arguments{
\item{pattern}{regular expression pattern applied on key bindings to filter shortcuts of interest}

\item{ignore_case}{TRUE or FALSE to ignore case in regex pattern search, TRUE by default}

\item{data}{shortcuts_table with tibble created based on RStudio Support page}
}
\value{
a tibble with matching observations
}
\description{
Finds shortcuts based on matched pattern on mac keybindings.
}
\examples{
find_kb_mac("Ctrl+1")
}
