% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ks.R
\name{find_ks}
\alias{find_ks}
\title{Finds matching keyboard shortcuts based on matched pattern on description}
\usage{
find_ks(
  pattern,
  ignore_case = TRUE,
  windows_linux_include = TRUE,
  mac_include = TRUE,
  type_include = TRUE,
  data = shortcuts_table
)
}
\arguments{
\item{pattern}{regular expression pattern applied on description to filter shortcuts of interest}

\item{ignore_case}{TRUE or FALSE to ignore case in regex pattern search, TRUE by default}

\item{windows_linux_include}{TRUE or FALSE to include column with windows/linux shortcuts, TRUE by default}

\item{mac_include}{TRUE or FALSE to include column with max shortcuts, TRUE by default}

\item{type_include}{TRUE or FALSE to include column with type info, TRUE by default}

\item{data}{shortcuts_table with tibble created based on RStudio Support page}
}
\value{
a tibble with matching observations
}
\description{
Finds matching keyboard shortcuts based on matched pattern on description
}
\examples{
find_ks("zoom console")
find_ks("terminal", windows_linux_include = FALSE)

}
