% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shortcuts_table}
\alias{shortcuts_table}
\title{RStudio IDE default keyboard shortcuts table.}
\format{
A tibble with one observation for each keyboard shortcut with the following columns
\describe{
\item{description}{description of shortcut}
\item{windows_linux}{windows/linux shortcut}
\item{mac}{mac keyboard shortcut}
\item{type}{type of shortcut}
}
}
\source{
\url{https://github.com/parmsam/keyboard-shortcuts-rstudio-ide/}

\url{https://support.rstudio.com/hc/en-us/articles/200711853-Keyboard-Shortcuts}
}
\usage{
shortcuts_table
}
\description{
A dataset containing default RStudio IDE Keyboard shortcuts for Windows, Linux, and Mac.
Dataset created based on RStudio Support page. See source for web link.
}
\keyword{datasets}
